backend/requirements.txt
Flask
Flask-Cors
python-dotenv
requests
openai
google-generativeai
services